# docs/swagger.yaml
definitions:
  handler.ConfigSetRequest:
    properties:
      appid:
        type: integer
      cache_size:
        type: integer
      enable_https:
        type: boolean
      max_download_load:
        type: integer
      max_parallel:
        type: integer
      max_upload_parallel:
        type: integer
      pan_ua:
        type: string
      pcs_ua:
        type: string
      user_agent:
        type: string
    type: object
  model.CloudAddRequest:
    properties:
      save_path:
        description: 保存路径
        type: string
      source_urls:
        description: 下载源地址列表
        items:
          type: string
        minItems: 1
        type: array
    required:
    - save_path
    - source_urls
    type: object
  model.CloudQueryRequest:
    properties:
      task_ids:
        description: 任务ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - task_ids
    type: object
  model.CloudTaskIDsRequest:
    properties:
      task_ids:
        description: 任务ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - task_ids
    type: object
...
  model.LoginRequest:
    properties:
      bduss:
        type: string
      password:
        type: string
      username:
        type: string
      vcode:
        type: string
      vcodestr:
        type: string
    type: object
...
  model.Response:
    properties:
      code:
        description: 状态码：0-成功，非0-失败
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
  model.UserSwitchRequest:
    properties:
      uid:
        type: integer
    required:
    - uid
    type: object
info:
  contact: {}
paths:
...
  /api/account/quota:
    get:
      consumes:
      - application/json
      description: 获取当前账号的空间配额和使用情况
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取网盘配额
      tags:
      - 账号管理
  /api/account/switch:
    post:
      consumes:
      - application/json
      description: 切换当前活跃的百度网盘账号
      parameters:
      - description: 切换用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserSwitchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 切换账号
      tags:
      - 账号管理
  /api/account/who:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取当前账号信息
      tags:
      - 账号管理
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 使用 BDUSS 或 用户名/密码 登录百度网盘
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      summary: 登录账号
      tags:
      - 账号管理
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: 退出当前登录的百度网盘账号
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: 退出当前账号
      tags:
      - 账号管理
...
  /api/cloud/add:
    post:
      consumes:
      - application/json
      description: 添加新的离线下载任务
      parameters:
      - description: 添加任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CloudAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 添加离线下载
      tags:
      - 离线下载
  /api/cloud/cancel:
    post:
      consumes:
      - application/json
      description: 取消正在进行的离线下载任务
      parameters:
      - description: 取消任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CloudTaskIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 取消离线任务
      tags:
      - 离线下载
  /api/cloud/clear:
    post:
      consumes:
      - application/json
      description: 清空所有离线下载任务记录
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 清空离线任务
      tags:
      - 离线下载
  /api/cloud/delete:
    post:
      consumes:
      - application/json
      description: 删除离线下载任务记录
      parameters:
      - description: 删除任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CloudTaskIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除离线任务
      tags:
      - 离线下载
  /api/cloud/list:
    get:
      consumes:
      - application/json
      description: 列出所有的离线下载任务
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 列出离线任务
      tags:
      - 离线下载
  /api/cloud/query:
    post:
      consumes:
      - application/json
      description: 查询指定 ID 的离线任务状态
      parameters:
      - description: 查询任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CloudQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 查询离线任务
      tags:
      - 离线下载
...
  /api/config/set:
    post:
      consumes:
      - application/json
      description: 修改 API 服务的配置项
      parameters:
      - description: 配置设置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ConfigSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 设置配置
      tags:
      - 配置管理
...
swagger: "2.0"